quote_verifier/requirements.txt
azure-security-attestation
python-dotenv
azure-identity
requests
